Compute shadow camera bounding box once
